[FE-183] style: inputTextarea 컴포넌트 분리
